+++ src/Main/ProgramSettings.h
@@ -125,15 +125,6 @@ struct ProgramSettings {
     void SetSettingsFromJSON(const nlohmann::json& j);
     void PushFilePaths();
 
-    // TODO: this is outdated. It's used by the equality operator, which is used by a test.
-    auto GetTuple() const {
-        return std::tie(help, version, port, omp_thread_count, top_level_folder, starting_folder, host, files, frontend_folder, no_http,
-            no_browser, no_log, log_performance, log_protocol_messages, debug_no_auth, controller_deployment, verbosity, wait_time,
-            init_wait_time, idle_session_wait_time, full_image_cache);
-    }
-    bool operator!=(const ProgramSettings& rhs) const;
-    bool operator==(const ProgramSettings& rhs) const;
-
     std::vector<std::string> warning_msgs;
     std::vector<std::string> debug_msgs;
     void FlushMessages() {