+++ src/ImageData/CompressedFits.h
@@ -4,8 +4,8 @@
    SPDX-License-Identifier: GPL-3.0-or-later
 */
 
-#ifndef CARTA_BACKEND_IMAGEDATA_COMPRESSEDFITS_H_
-#define CARTA_BACKEND_IMAGEDATA_COMPRESSEDFITS_H_
+#ifndef CARTA_SRC_IMAGEDATA_COMPRESSEDFITS_H_
+#define CARTA_SRC_IMAGEDATA_COMPRESSEDFITS_H_
 
 #include <zlib.h>
 #include <map>
@@ -159,4 +159,4 @@ private:
 
 } // namespace carta
 
-#endif // CARTA_BACKEND_IMAGEDATA_COMPRESSEDFITS_H_
+#endif // CARTA_SRC_IMAGEDATA_COMPRESSEDFITS_H_