+++ src/ImageData/ExprLoader.h
@@ -4,8 +4,8 @@
    SPDX-License-Identifier: GPL-3.0-or-later
 */
 
-#ifndef CARTA_BACKEND_IMAGEDATA_EXPRLOADER_H_
-#define CARTA_BACKEND_IMAGEDATA_EXPRLOADER_H_
+#ifndef CARTA_SRC_IMAGEDATA_EXPRLOADER_H_
+#define CARTA_SRC_IMAGEDATA_EXPRLOADER_H_
 
 #include <casacore/casa/Json/JsonKVMap.h>
 #include <casacore/casa/Json/JsonParser.h>
@@ -93,4 +93,4 @@ bool ExprLoader::SaveFile(const CARTA::FileType type, const std::string& output_
 
 } // namespace carta
 
-#endif // CARTA_BACKEND_IMAGEDATA_EXPRLOADER_H_
+#endif // CARTA_SRC_IMAGEDATA_EXPRLOADER_H_