+++ src/Logger/Logger.h
@@ -4,8 +4,8 @@
    SPDX-License-Identifier: GPL-3.0-or-later
 */
 
-#ifndef CARTA_BACKEND_LOGGER_LOGGER_H_
-#define CARTA_BACKEND_LOGGER_LOGGER_H_
+#ifndef CARTA_SRC_LOGGER_LOGGER_H_
+#define CARTA_SRC_LOGGER_LOGGER_H_
 
 #include <iostream>
 
@@ -119,4 +119,4 @@ void FlushLogFile();
 } // namespace logger
 } // namespace carta
 
-#endif // CARTA_BACKEND_LOGGER_LOGGER_H_
+#endif // CARTA_SRC_LOGGER_LOGGER_H_