+++ src/Main/ProgramSettings.h
@@ -4,8 +4,8 @@
    SPDX-License-Identifier: GPL-3.0-or-later
 */
 
-#ifndef CARTA_BACKEND_SRC_MAIN_PROGRAMSETTINGS_H_
-#define CARTA_BACKEND_SRC_MAIN_PROGRAMSETTINGS_H_
+#ifndef CARTA_SRC_MAIN_PROGRAMSETTINGS_H_
+#define CARTA_SRC_MAIN_PROGRAMSETTINGS_H_
 
 #include <iostream>
 #include <string>
@@ -141,4 +141,4 @@ struct ProgramSettings {
     }
 };
 } // namespace carta
-#endif // CARTA_BACKEND_SRC_MAIN_PROGRAMSETTINGS_H_
+#endif // CARTA_SRC_MAIN_PROGRAMSETTINGS_H_