+++ src/Region/Region.cc
@@ -163,7 +163,8 @@ void Region::SetReferenceRegion() {
                 if (ConvertCartaPointToWorld(pixel_points[0], _wcs_control_points)) {
                     // WCBox blc and trc are same point
                     std::lock_guard<std::mutex> guard(_region_mutex);
-                    region = new casacore::WCBox(_wcs_control_points, _wcs_control_points, pixel_axes, *_coord_sys, abs_rel);
+                    casacore::Vector<casacore::Quantity> const wcs_points(_wcs_control_points);
+                    region = new casacore::WCBox(wcs_points, wcs_points, pixel_axes, *_coord_sys, abs_rel);
                 }
                 break;
             }